feat(admin): add mailing lists opt-out and data quota settings pages to service account overview * added page for mailing lists opt-out for service identities * added page for data quota management for service identities * refactored some components
