+++ apps/admin-gui/src/app/users/pages/user-detail-page/user-settings/user-settings-mailing-lists/user-settings-mailing-lists.component.html
@@ -1,0 +1,1 @@
+<perun-web-apps-mailing-lists [user]="user" [isService]="true"></perun-web-apps-mailing-lists>
