+++ apps/admin-gui/src/app/users/pages/user-detail-page/user-settings/user-settings-service-identities/service-identity-detail-page/service-identity-overview/service-identity-overview.component.ts
@@ -27,6 +27,18 @@ export class ServiceIdentityOverviewComponent implements OnInit {
         label: 'MENU_ITEMS.USER.AUTHENTICATION',
         style: 'user-btn',
       },
+      {
+        cssIcon: 'perun-notification',
+        url: 'mailing-lists',
+        label: 'MENU_ITEMS.USER.MAILING_LISTS',
+        style: 'user-btn',
+      },
+      {
+        cssIcon: 'perun-statistics',
+        url: 'data-quotas',
+        label: 'MENU_ITEMS.USER.DATA_QUOTAS',
+        style: 'user-btn',
+      },
     ];
   }
 }
