+++ apps/user-profile/src/app/pages/settings-page/settings-mailing-lists/settings-mailing-lists.component.html
@@ -1,45 +1,1 @@
-<h1 class="page-title">{{'OPT_OUT_MAILING_LISTS.TITLE' | customTranslate | translate}}</h1>
-<perun-web-apps-immediate-filter
-  [placeholder]="'OPT_OUT_MAILING_LISTS.FILTER' | customTranslate | translate"
-  (filter)="applyFilter($event)"></perun-web-apps-immediate-filter>
-<mat-accordion>
-  <mat-expansion-panel
-    *ngFor="let vo of filteredVos"
-    (closed)="deselectVo()"
-    (afterExpand)="changeSelectedVo(vo)"
-    [expanded]="vo.shortName === selectedVo">
-    <mat-expansion-panel-header>
-      <mat-panel-title>
-        {{vo.name}}
-      </mat-panel-title>
-    </mat-expansion-panel-header>
-    <ng-template matExpansionPanelContent>
-      <mat-spinner class="ml-auto mr-auto" *ngIf="loading"></mat-spinner>
-      <perun-web-apps-alert
-        *ngIf="resources.length === 0 && !loading"
-        alert_type="warn"
-        >{{'OPT_OUT_MAILING_LISTS.NO_MAILING_LISTS_ALERT' | customTranslate | translate}}</perun-web-apps-alert
-      >
-      <mat-accordion *ngIf="resources.length && !loading">
-        <mat-expansion-panel
-          *ngFor="let resource of resources"
-          (closed)="deselectResource()"
-          (afterExpand)="changeSelectedResource(resource)"
-          [expanded]="resource.name === selectedResource">
-          <mat-expansion-panel-header>
-            <mat-panel-title>
-              {{resource.name}}
-            </mat-panel-title>
-          </mat-expansion-panel-header>
-          <div class="row">
-            <mat-checkbox
-              class="ml-3 mr-2"
-              [checked]="optOutAttribute && optOutAttribute.value !== null"
-              (change)="setOptOut()"></mat-checkbox>
-            <p>{{'OPT_OUT_MAILING_LISTS.OPT_OUT_LABEL' | customTranslate | translate}}</p>
-          </div>
-        </mat-expansion-panel>
-      </mat-accordion>
-    </ng-template>
-  </mat-expansion-panel>
-</mat-accordion>
+<perun-web-apps-mailing-lists [user]="user" [isService]="false"></perun-web-apps-mailing-lists>
