+++ libs/perun/components/src/lib/perun-components.module.ts
@@ -99,6 +99,8 @@ import { AttributeRightsItemComponent } from './attribute-rights-item/attribute-
 import { AttributeRightsTabGroupComponent } from './attribute-rights-tab-group/attribute-rights-tab-group.component';
 import { AttributeUniqueToggleComponent } from './attribute-unique-toggle/attribute-unique-toggle.component';
 import { AttributeCriticalOperationsTogglesComponent } from './attribute-critical-operations-toggles/attribute-critical-operations-toggles.component';
+import { MailingListsComponent } from './settings-mailing-lists/mailing-lists.component';
+import { DataQuotasComponent } from './settings-data-quotas/data-quotas.component';
 import { LoaderDirective } from '@perun-web-apps/perun/directives';
 
 @Injectable()
@@ -232,6 +234,8 @@ export const APP_DATE_FORMATS: MatDateFormats = {
     AttributeUniqueToggleComponent,
     AttributeCriticalOperationsTogglesComponent,
     LoaderDirective,
+    MailingListsComponent,
+    DataQuotasComponent,
   ],
   exports: [
     VosListComponent,
@@ -292,6 +296,8 @@ export const APP_DATE_FORMATS: MatDateFormats = {
     AttributeUniqueToggleComponent,
     AttributeCriticalOperationsTogglesComponent,
     LoaderDirective,
+    MailingListsComponent,
+    DataQuotasComponent,
   ],
   providers: [
     { provide: DateAdapter, useClass: AppDateAdapter },
