+++ libs/perun/components/src/lib/settings-data-quotas/data-quotas.component.scss
@@ -1,0 +1,3 @@
+.color-gray {
+  color: gray;
+}
